move simplifier to expr_simplifier
